+++ packages/wbw_core/lib/src/models/models.freezed.dart
@@ -20,7 +20,7 @@ AppSettingsModel _$AppSettingsModelFromJson(Map<String, dynamic> json) {
 
 /// @nodoc
 mixin _$AppSettingsModel {
-  @JsonKey(fromJson: _localeFromString, toJson: _localeToString)
+  @JsonKey(fromJson: localeFromString, toJson: localeToString)
   Locale get locale => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
@@ -35,7 +35,7 @@ abstract class $AppSettingsModelCopyWith<$Res> {
           AppSettingsModel value, $Res Function(AppSettingsModel) then) =
       _$AppSettingsModelCopyWithImpl<$Res>;
   $Res call(
-      {@JsonKey(fromJson: _localeFromString, toJson: _localeToString)
+      {@JsonKey(fromJson: localeFromString, toJson: localeToString)
           Locale locale});
 }
 
@@ -69,7 +69,7 @@ abstract class _$$_AppSettingsModelCopyWith<$Res>
       __$$_AppSettingsModelCopyWithImpl<$Res>;
   @override
   $Res call(
-      {@JsonKey(fromJson: _localeFromString, toJson: _localeToString)
+      {@JsonKey(fromJson: localeFromString, toJson: localeToString)
           Locale locale});
 }
 
@@ -102,7 +102,7 @@ class __$$_AppSettingsModelCopyWithImpl<$Res>
 @JsonSerializable(explicitToJson: true)
 class _$_AppSettingsModel extends _AppSettingsModel {
   const _$_AppSettingsModel(
-      {@JsonKey(fromJson: _localeFromString, toJson: _localeToString)
+      {@JsonKey(fromJson: localeFromString, toJson: localeToString)
           this.locale = Locales.en})
       : super._();
 
@@ -110,7 +110,7 @@ class _$_AppSettingsModel extends _AppSettingsModel {
       _$$_AppSettingsModelFromJson(json);
 
   @override
-  @JsonKey(fromJson: _localeFromString, toJson: _localeToString)
+  @JsonKey(fromJson: localeFromString, toJson: localeToString)
   final Locale locale;
 
   @override
@@ -146,7 +146,7 @@ class _$_AppSettingsModel extends _AppSettingsModel {
 
 abstract class _AppSettingsModel extends AppSettingsModel {
   const factory _AppSettingsModel(
-      {@JsonKey(fromJson: _localeFromString, toJson: _localeToString)
+      {@JsonKey(fromJson: localeFromString, toJson: localeToString)
           final Locale locale}) = _$_AppSettingsModel;
   const _AppSettingsModel._() : super._();
 
@@ -154,7 +154,7 @@ abstract class _AppSettingsModel extends AppSettingsModel {
       _$_AppSettingsModel.fromJson;
 
   @override
-  @JsonKey(fromJson: _localeFromString, toJson: _localeToString)
+  @JsonKey(fromJson: localeFromString, toJson: localeToString)
   Locale get locale;
   @override
   @JsonKey(ignore: true)
@@ -1225,12 +1225,11 @@ LevelModel _$LevelModelFromJson(Map<String, dynamic> json) {
 
 /// @nodoc
 mixin _$LevelModel {
-  String get id =>
-      throw _privateConstructorUsedError; // TODO(arenukvern): add i18n object
-  String get stringName => throw _privateConstructorUsedError;
+  String get id => throw _privateConstructorUsedError;
   LevelPlayersModel get players => throw _privateConstructorUsedError;
   LevelCharactersModel get characters => throw _privateConstructorUsedError;
   ResourcesModel get resources => throw _privateConstructorUsedError;
+  LocalizedMap get name => throw _privateConstructorUsedError;
   CurrentWordModel get currentWord => throw _privateConstructorUsedError;
   Map<String, String> get words => throw _privateConstructorUsedError;
   String get latestWord => throw _privateConstructorUsedError;
@@ -1248,10 +1247,10 @@ abstract class $LevelModelCopyWith<$Res> {
       _$LevelModelCopyWithImpl<$Res>;
   $Res call(
       {String id,
-      String stringName,
       LevelPlayersModel players,
       LevelCharactersModel characters,
       ResourcesModel resources,
+      LocalizedMap name,
       CurrentWordModel currentWord,
       Map<String, String> words,
       String latestWord});
@@ -1259,6 +1258,7 @@ abstract class $LevelModelCopyWith<$Res> {
   $LevelPlayersModelCopyWith<$Res> get players;
   $LevelCharactersModelCopyWith<$Res> get characters;
   $ResourcesModelCopyWith<$Res> get resources;
+  $LocalizedMapCopyWith<$Res> get name;
   $CurrentWordModelCopyWith<$Res> get currentWord;
 }
 
@@ -1273,10 +1273,10 @@ class _$LevelModelCopyWithImpl<$Res> implements $LevelModelCopyWith<$Res> {
   @override
   $Res call({
     Object? id = freezed,
-    Object? stringName = freezed,
     Object? players = freezed,
     Object? characters = freezed,
     Object? resources = freezed,
+    Object? name = freezed,
     Object? currentWord = freezed,
     Object? words = freezed,
     Object? latestWord = freezed,
@@ -1285,10 +1285,6 @@ class _$LevelModelCopyWithImpl<$Res> implements $LevelModelCopyWith<$Res> {
       id: id == freezed
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
-              as String,
-      stringName: stringName == freezed
-          ? _value.stringName
-          : stringName // ignore: cast_nullable_to_non_nullable
               as String,
       players: players == freezed
           ? _value.players
@@ -1302,6 +1298,10 @@ class _$LevelModelCopyWithImpl<$Res> implements $LevelModelCopyWith<$Res> {
           ? _value.resources
           : resources // ignore: cast_nullable_to_non_nullable
               as ResourcesModel,
+      name: name == freezed
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as LocalizedMap,
       currentWord: currentWord == freezed
           ? _value.currentWord
           : currentWord // ignore: cast_nullable_to_non_nullable
@@ -1339,6 +1339,13 @@ class _$LevelModelCopyWithImpl<$Res> implements $LevelModelCopyWith<$Res> {
   }
 
   @override
+  $LocalizedMapCopyWith<$Res> get name {
+    return $LocalizedMapCopyWith<$Res>(_value.name, (value) {
+      return _then(_value.copyWith(name: value));
+    });
+  }
+
+  @override
   $CurrentWordModelCopyWith<$Res> get currentWord {
     return $CurrentWordModelCopyWith<$Res>(_value.currentWord, (value) {
       return _then(_value.copyWith(currentWord: value));
@@ -1355,10 +1362,10 @@ abstract class _$$_LevelModelCopyWith<$Res>
   @override
   $Res call(
       {String id,
-      String stringName,
       LevelPlayersModel players,
       LevelCharactersModel characters,
       ResourcesModel resources,
+      LocalizedMap name,
       CurrentWordModel currentWord,
       Map<String, String> words,
       String latestWord});
@@ -1369,6 +1376,8 @@ abstract class _$$_LevelModelCopyWith<$Res>
   $LevelCharactersModelCopyWith<$Res> get characters;
   @override
   $ResourcesModelCopyWith<$Res> get resources;
+  @override
+  $LocalizedMapCopyWith<$Res> get name;
   @override
   $CurrentWordModelCopyWith<$Res> get currentWord;
 }
@@ -1386,10 +1395,10 @@ class __$$_LevelModelCopyWithImpl<$Res> extends _$LevelModelCopyWithImpl<$Res>
   @override
   $Res call({
     Object? id = freezed,
-    Object? stringName = freezed,
     Object? players = freezed,
     Object? characters = freezed,
     Object? resources = freezed,
+    Object? name = freezed,
     Object? currentWord = freezed,
     Object? words = freezed,
     Object? latestWord = freezed,
@@ -1398,10 +1407,6 @@ class __$$_LevelModelCopyWithImpl<$Res> extends _$LevelModelCopyWithImpl<$Res>
       id: id == freezed
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
-              as String,
-      stringName: stringName == freezed
-          ? _value.stringName
-          : stringName // ignore: cast_nullable_to_non_nullable
               as String,
       players: players == freezed
           ? _value.players
@@ -1415,6 +1420,10 @@ class __$$_LevelModelCopyWithImpl<$Res> extends _$LevelModelCopyWithImpl<$Res>
           ? _value.resources
           : resources // ignore: cast_nullable_to_non_nullable
               as ResourcesModel,
+      name: name == freezed
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as LocalizedMap,
       currentWord: currentWord == freezed
           ? _value.currentWord
           : currentWord // ignore: cast_nullable_to_non_nullable
@@ -1437,10 +1446,10 @@ class __$$_LevelModelCopyWithImpl<$Res> extends _$LevelModelCopyWithImpl<$Res>
 class _$_LevelModel extends _LevelModel {
   const _$_LevelModel(
       {required this.id,
-      required this.stringName,
       required this.players,
       required this.characters,
       required this.resources,
+      this.name = LocalizedMap.empty,
       this.currentWord = const CurrentWordModel(),
       final Map<String, String> words = const {},
       this.latestWord = ''})
@@ -1452,15 +1461,15 @@ class _$_LevelModel extends _LevelModel {
 
   @override
   final String id;
-// TODO(arenukvern): add i18n object
-  @override
-  final String stringName;
   @override
   final LevelPlayersModel players;
   @override
   final LevelCharactersModel characters;
   @override
   final ResourcesModel resources;
+  @override
+  @JsonKey()
+  final LocalizedMap name;
   @override
   @JsonKey()
   final CurrentWordModel currentWord;
@@ -1478,7 +1487,7 @@ class _$_LevelModel extends _LevelModel {
 
   @override
   String toString() {
-    return 'LevelModel(id: $id, stringName: $stringName, players: $players, characters: $characters, resources: $resources, currentWord: $currentWord, words: $words, latestWord: $latestWord)';
+    return 'LevelModel(id: $id, players: $players, characters: $characters, resources: $resources, name: $name, currentWord: $currentWord, words: $words, latestWord: $latestWord)';
   }
 
   @override
@@ -1487,12 +1496,11 @@ class _$_LevelModel extends _LevelModel {
         (other.runtimeType == runtimeType &&
             other is _$_LevelModel &&
             const DeepCollectionEquality().equals(other.id, id) &&
-            const DeepCollectionEquality()
-                .equals(other.stringName, stringName) &&
             const DeepCollectionEquality().equals(other.players, players) &&
             const DeepCollectionEquality()
                 .equals(other.characters, characters) &&
             const DeepCollectionEquality().equals(other.resources, resources) &&
+            const DeepCollectionEquality().equals(other.name, name) &&
             const DeepCollectionEquality()
                 .equals(other.currentWord, currentWord) &&
             const DeepCollectionEquality().equals(other._words, _words) &&
@@ -1505,10 +1513,10 @@ class _$_LevelModel extends _LevelModel {
   int get hashCode => Object.hash(
       runtimeType,
       const DeepCollectionEquality().hash(id),
-      const DeepCollectionEquality().hash(stringName),
       const DeepCollectionEquality().hash(players),
       const DeepCollectionEquality().hash(characters),
       const DeepCollectionEquality().hash(resources),
+      const DeepCollectionEquality().hash(name),
       const DeepCollectionEquality().hash(currentWord),
       const DeepCollectionEquality().hash(_words),
       const DeepCollectionEquality().hash(latestWord));
@@ -1529,10 +1537,10 @@ class _$_LevelModel extends _LevelModel {
 abstract class _LevelModel extends LevelModel {
   const factory _LevelModel(
       {required final String id,
-      required final String stringName,
       required final LevelPlayersModel players,
       required final LevelCharactersModel characters,
       required final ResourcesModel resources,
+      final LocalizedMap name,
       final CurrentWordModel currentWord,
       final Map<String, String> words,
       final String latestWord}) = _$_LevelModel;
@@ -1543,14 +1551,14 @@ abstract class _LevelModel extends LevelModel {
 
   @override
   String get id;
-  @override // TODO(arenukvern): add i18n object
-  String get stringName;
   @override
   LevelPlayersModel get players;
   @override
   LevelCharactersModel get characters;
   @override
   ResourcesModel get resources;
+  @override
+  LocalizedMap get name;
   @override
   CurrentWordModel get currentWord;
   @override
@@ -1569,10 +1577,9 @@ TemplateLevelModel _$TemplateLevelModelFromJson(Map<String, dynamic> json) {
 
 /// @nodoc
 mixin _$TemplateLevelModel {
-  String get id =>
-      throw _privateConstructorUsedError; // TODO(arenukvern): add i18n object
-  String get stringName => throw _privateConstructorUsedError;
+  String get id => throw _privateConstructorUsedError;
   ResourcesModel get resources => throw _privateConstructorUsedError;
+  LocalizedMap get name => throw _privateConstructorUsedError;
   FuelStorageModel get fuelStorage => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
@@ -1588,11 +1595,12 @@ abstract class $TemplateLevelModelCopyWith<$Res> {
       _$TemplateLevelModelCopyWithImpl<$Res>;
   $Res call(
       {String id,
-      String stringName,
       ResourcesModel resources,
+      LocalizedMap name,
       FuelStorageModel fuelStorage});
 
   $ResourcesModelCopyWith<$Res> get resources;
+  $LocalizedMapCopyWith<$Res> get name;
   $FuelStorageModelCopyWith<$Res> get fuelStorage;
 }
 
@@ -1608,8 +1616,8 @@ class _$TemplateLevelModelCopyWithImpl<$Res>
   @override
   $Res call({
     Object? id = freezed,
-    Object? stringName = freezed,
     Object? resources = freezed,
+    Object? name = freezed,
     Object? fuelStorage = freezed,
   }) {
     return _then(_value.copyWith(
@@ -1617,14 +1625,14 @@ class _$TemplateLevelModelCopyWithImpl<$Res>
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
               as String,
-      stringName: stringName == freezed
-          ? _value.stringName
-          : stringName // ignore: cast_nullable_to_non_nullable
-              as String,
       resources: resources == freezed
           ? _value.resources
           : resources // ignore: cast_nullable_to_non_nullable
               as ResourcesModel,
+      name: name == freezed
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as LocalizedMap,
       fuelStorage: fuelStorage == freezed
           ? _value.fuelStorage
           : fuelStorage // ignore: cast_nullable_to_non_nullable
@@ -1636,6 +1644,13 @@ class _$TemplateLevelModelCopyWithImpl<$Res>
   $ResourcesModelCopyWith<$Res> get resources {
     return $ResourcesModelCopyWith<$Res>(_value.resources, (value) {
       return _then(_value.copyWith(resources: value));
+    });
+  }
+
+  @override
+  $LocalizedMapCopyWith<$Res> get name {
+    return $LocalizedMapCopyWith<$Res>(_value.name, (value) {
+      return _then(_value.copyWith(name: value));
     });
   }
 
@@ -1656,12 +1671,14 @@ abstract class _$$_TemplateLevelModelCopyWith<$Res>
   @override
   $Res call(
       {String id,
-      String stringName,
       ResourcesModel resources,
+      LocalizedMap name,
       FuelStorageModel fuelStorage});
 
   @override
   $ResourcesModelCopyWith<$Res> get resources;
+  @override
+  $LocalizedMapCopyWith<$Res> get name;
   @override
   $FuelStorageModelCopyWith<$Res> get fuelStorage;
 }
@@ -1680,8 +1697,8 @@ class __$$_TemplateLevelModelCopyWithImpl<$Res>
   @override
   $Res call({
     Object? id = freezed,
-    Object? stringName = freezed,
     Object? resources = freezed,
+    Object? name = freezed,
     Object? fuelStorage = freezed,
   }) {
     return _then(_$_TemplateLevelModel(
@@ -1689,14 +1706,14 @@ class __$$_TemplateLevelModelCopyWithImpl<$Res>
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
               as String,
-      stringName: stringName == freezed
-          ? _value.stringName
-          : stringName // ignore: cast_nullable_to_non_nullable
-              as String,
       resources: resources == freezed
           ? _value.resources
           : resources // ignore: cast_nullable_to_non_nullable
               as ResourcesModel,
+      name: name == freezed
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as LocalizedMap,
       fuelStorage: fuelStorage == freezed
           ? _value.fuelStorage
           : fuelStorage // ignore: cast_nullable_to_non_nullable
@@ -1711,8 +1728,8 @@ class __$$_TemplateLevelModelCopyWithImpl<$Res>
 class _$_TemplateLevelModel extends _TemplateLevelModel {
   const _$_TemplateLevelModel(
       {required this.id,
-      required this.stringName,
       required this.resources,
+      this.name = LocalizedMap.empty,
       this.fuelStorage = const FuelStorageModel()})
       : super._();
 
@@ -1721,18 +1738,18 @@ class _$_TemplateLevelModel extends _TemplateLevelModel {
 
   @override
   final String id;
-// TODO(arenukvern): add i18n object
-  @override
-  final String stringName;
   @override
   final ResourcesModel resources;
+  @override
+  @JsonKey()
+  final LocalizedMap name;
   @override
   @JsonKey()
   final FuelStorageModel fuelStorage;
 
   @override
   String toString() {
-    return 'TemplateLevelModel(id: $id, stringName: $stringName, resources: $resources, fuelStorage: $fuelStorage)';
+    return 'TemplateLevelModel(id: $id, resources: $resources, name: $name, fuelStorage: $fuelStorage)';
   }
 
   @override
@@ -1741,9 +1758,8 @@ class _$_TemplateLevelModel extends _TemplateLevelModel {
         (other.runtimeType == runtimeType &&
             other is _$_TemplateLevelModel &&
             const DeepCollectionEquality().equals(other.id, id) &&
-            const DeepCollectionEquality()
-                .equals(other.stringName, stringName) &&
             const DeepCollectionEquality().equals(other.resources, resources) &&
+            const DeepCollectionEquality().equals(other.name, name) &&
             const DeepCollectionEquality()
                 .equals(other.fuelStorage, fuelStorage));
   }
@@ -1753,8 +1769,8 @@ class _$_TemplateLevelModel extends _TemplateLevelModel {
   int get hashCode => Object.hash(
       runtimeType,
       const DeepCollectionEquality().hash(id),
-      const DeepCollectionEquality().hash(stringName),
       const DeepCollectionEquality().hash(resources),
+      const DeepCollectionEquality().hash(name),
       const DeepCollectionEquality().hash(fuelStorage));
 
   @JsonKey(ignore: true)
@@ -1774,8 +1790,8 @@ class _$_TemplateLevelModel extends _TemplateLevelModel {
 abstract class _TemplateLevelModel extends TemplateLevelModel {
   const factory _TemplateLevelModel(
       {required final String id,
-      required final String stringName,
       required final ResourcesModel resources,
+      final LocalizedMap name,
       final FuelStorageModel fuelStorage}) = _$_TemplateLevelModel;
   const _TemplateLevelModel._() : super._();
 
@@ -1784,10 +1800,10 @@ abstract class _TemplateLevelModel extends TemplateLevelModel {
 
   @override
   String get id;
-  @override // TODO(arenukvern): add i18n object
-  String get stringName;
   @override
   ResourcesModel get resources;
+  @override
+  LocalizedMap get name;
   @override
   FuelStorageModel get fuelStorage;
   @override
@@ -2113,10 +2129,10 @@ PlayerCharacterModel _$PlayerCharacterModelFromJson(Map<String, dynamic> json) {
 /// @nodoc
 mixin _$PlayerCharacterModel {
   String get id => throw _privateConstructorUsedError;
-  String get name => throw _privateConstructorUsedError;
   String get description => throw _privateConstructorUsedError;
   int get color => throw _privateConstructorUsedError;
   CharacterAssetModel get asset => throw _privateConstructorUsedError;
+  LocalizedMap get localizedName => throw _privateConstructorUsedError;
   String get characterIcon => throw _privateConstructorUsedError;
   SerializedVector2 get position => throw _privateConstructorUsedError;
   FuelStorageModel get fuel => throw _privateConstructorUsedError;
@@ -2136,10 +2152,10 @@ abstract class $PlayerCharacterModelCopyWith<$Res> {
       _$PlayerCharacterModelCopyWithImpl<$Res>;
   $Res call(
       {String id,
-      String name,
       String description,
       int color,
       CharacterAssetModel asset,
+      LocalizedMap localizedName,
       String characterIcon,
       SerializedVector2 position,
       FuelStorageModel fuel,
@@ -2147,6 +2163,7 @@ abstract class $PlayerCharacterModelCopyWith<$Res> {
       double requiredLiftForce});
 
   $CharacterAssetModelCopyWith<$Res> get asset;
+  $LocalizedMapCopyWith<$Res> get localizedName;
   $SerializedVector2CopyWith<$Res> get position;
   $FuelStorageModelCopyWith<$Res> get fuel;
 }
@@ -2163,10 +2180,10 @@ class _$PlayerCharacterModelCopyWithImpl<$Res>
   @override
   $Res call({
     Object? id = freezed,
-    Object? name = freezed,
     Object? description = freezed,
     Object? color = freezed,
     Object? asset = freezed,
+    Object? localizedName = freezed,
     Object? characterIcon = freezed,
     Object? position = freezed,
     Object? fuel = freezed,
@@ -2177,10 +2194,6 @@ class _$PlayerCharacterModelCopyWithImpl<$Res>
       id: id == freezed
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
-              as String,
-      name: name == freezed
-          ? _value.name
-          : name // ignore: cast_nullable_to_non_nullable
               as String,
       description: description == freezed
           ? _value.description
@@ -2194,6 +2207,10 @@ class _$PlayerCharacterModelCopyWithImpl<$Res>
           ? _value.asset
           : asset // ignore: cast_nullable_to_non_nullable
               as CharacterAssetModel,
+      localizedName: localizedName == freezed
+          ? _value.localizedName
+          : localizedName // ignore: cast_nullable_to_non_nullable
+              as LocalizedMap,
       characterIcon: characterIcon == freezed
           ? _value.characterIcon
           : characterIcon // ignore: cast_nullable_to_non_nullable
@@ -2225,6 +2242,13 @@ class _$PlayerCharacterModelCopyWithImpl<$Res>
   }
 
   @override
+  $LocalizedMapCopyWith<$Res> get localizedName {
+    return $LocalizedMapCopyWith<$Res>(_value.localizedName, (value) {
+      return _then(_value.copyWith(localizedName: value));
+    });
+  }
+
+  @override
   $SerializedVector2CopyWith<$Res> get position {
     return $SerializedVector2CopyWith<$Res>(_value.position, (value) {
       return _then(_value.copyWith(position: value));
@@ -2248,10 +2272,10 @@ abstract class _$$_PlayerCharacterModelCopyWith<$Res>
   @override
   $Res call(
       {String id,
-      String name,
       String description,
       int color,
       CharacterAssetModel asset,
+      LocalizedMap localizedName,
       String characterIcon,
       SerializedVector2 position,
       FuelStorageModel fuel,
@@ -2260,6 +2284,8 @@ abstract class _$$_PlayerCharacterModelCopyWith<$Res>
 
   @override
   $CharacterAssetModelCopyWith<$Res> get asset;
+  @override
+  $LocalizedMapCopyWith<$Res> get localizedName;
   @override
   $SerializedVector2CopyWith<$Res> get position;
   @override
@@ -2280,10 +2306,10 @@ class __$$_PlayerCharacterModelCopyWithImpl<$Res>
   @override
   $Res call({
     Object? id = freezed,
-    Object? name = freezed,
     Object? description = freezed,
     Object? color = freezed,
     Object? asset = freezed,
+    Object? localizedName = freezed,
     Object? characterIcon = freezed,
     Object? position = freezed,
     Object? fuel = freezed,
@@ -2294,10 +2320,6 @@ class __$$_PlayerCharacterModelCopyWithImpl<$Res>
       id: id == freezed
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
-              as String,
-      name: name == freezed
-          ? _value.name
-          : name // ignore: cast_nullable_to_non_nullable
               as String,
       description: description == freezed
           ? _value.description
@@ -2311,6 +2333,10 @@ class __$$_PlayerCharacterModelCopyWithImpl<$Res>
           ? _value.asset
           : asset // ignore: cast_nullable_to_non_nullable
               as CharacterAssetModel,
+      localizedName: localizedName == freezed
+          ? _value.localizedName
+          : localizedName // ignore: cast_nullable_to_non_nullable
+              as LocalizedMap,
       characterIcon: characterIcon == freezed
           ? _value.characterIcon
           : characterIcon // ignore: cast_nullable_to_non_nullable
@@ -2341,10 +2367,10 @@ class __$$_PlayerCharacterModelCopyWithImpl<$Res>
 class _$_PlayerCharacterModel extends _PlayerCharacterModel {
   const _$_PlayerCharacterModel(
       {required this.id,
-      required this.name,
       required this.description,
       required this.color,
       required this.asset,
+      this.localizedName = LocalizedMap.empty,
       this.characterIcon = '',
       this.position = SerializedVector2.zero,
       this.fuel = const FuelStorageModel(value: 150),
@@ -2358,13 +2384,14 @@ class _$_PlayerCharacterModel extends _PlayerCharacterModel {
   @override
   final String id;
   @override
-  final String name;
-  @override
   final String description;
   @override
   final int color;
   @override
   final CharacterAssetModel asset;
+  @override
+  @JsonKey()
+  final LocalizedMap localizedName;
   @override
   @JsonKey()
   final String characterIcon;
@@ -2383,7 +2410,7 @@ class _$_PlayerCharacterModel extends _PlayerCharacterModel {
 
   @override
   String toString() {
-    return 'PlayerCharacterModel(id: $id, name: $name, description: $description, color: $color, asset: $asset, characterIcon: $characterIcon, position: $position, fuel: $fuel, fuelNormalPower: $fuelNormalPower, requiredLiftForce: $requiredLiftForce)';
+    return 'PlayerCharacterModel(id: $id, description: $description, color: $color, asset: $asset, localizedName: $localizedName, characterIcon: $characterIcon, position: $position, fuel: $fuel, fuelNormalPower: $fuelNormalPower, requiredLiftForce: $requiredLiftForce)';
   }
 
   @override
@@ -2392,11 +2419,12 @@ class _$_PlayerCharacterModel extends _PlayerCharacterModel {
         (other.runtimeType == runtimeType &&
             other is _$_PlayerCharacterModel &&
             const DeepCollectionEquality().equals(other.id, id) &&
-            const DeepCollectionEquality().equals(other.name, name) &&
             const DeepCollectionEquality()
                 .equals(other.description, description) &&
             const DeepCollectionEquality().equals(other.color, color) &&
             const DeepCollectionEquality().equals(other.asset, asset) &&
+            const DeepCollectionEquality()
+                .equals(other.localizedName, localizedName) &&
             const DeepCollectionEquality()
                 .equals(other.characterIcon, characterIcon) &&
             const DeepCollectionEquality().equals(other.position, position) &&
@@ -2412,10 +2440,10 @@ class _$_PlayerCharacterModel extends _PlayerCharacterModel {
   int get hashCode => Object.hash(
       runtimeType,
       const DeepCollectionEquality().hash(id),
-      const DeepCollectionEquality().hash(name),
       const DeepCollectionEquality().hash(description),
       const DeepCollectionEquality().hash(color),
       const DeepCollectionEquality().hash(asset),
+      const DeepCollectionEquality().hash(localizedName),
       const DeepCollectionEquality().hash(characterIcon),
       const DeepCollectionEquality().hash(position),
       const DeepCollectionEquality().hash(fuel),
@@ -2439,10 +2467,10 @@ class _$_PlayerCharacterModel extends _PlayerCharacterModel {
 abstract class _PlayerCharacterModel extends PlayerCharacterModel {
   const factory _PlayerCharacterModel(
       {required final String id,
-      required final String name,
       required final String description,
       required final int color,
       required final CharacterAssetModel asset,
+      final LocalizedMap localizedName,
       final String characterIcon,
       final SerializedVector2 position,
       final FuelStorageModel fuel,
@@ -2456,13 +2484,13 @@ abstract class _PlayerCharacterModel extends PlayerCharacterModel {
   @override
   String get id;
   @override
-  String get name;
-  @override
   String get description;
   @override
   int get color;
   @override
   CharacterAssetModel get asset;
+  @override
+  LocalizedMap get localizedName;
   @override
   String get characterIcon;
   @override